+++ game/skillribbon.py
@@ -1,0 +1,11 @@
+import pbge
+
+class SkillButton(pbge.widgets.ColumnWidget):
+    pass
+
+
+
+class SkillRibbon(pbge.widgets.Widget):
+    def __init__(self, pc):
+        pass
+